[chore] Exclude HttpClientFactory from PITest coverage in pom.xml
This commit updates the pom.xml to configure PITest to ignore the `HttpClientFactory` class during mutation testing. The exclusion of `HttpClientFactory` from PITest coverage is due to its nature as a configuration or infrastructure class, which may not benefit significantly from the mutation testing process. This adjustment helps to focus the mutation testing on more critical parts of the codebase, optimizing the testing efforts and resources.