[ci] Update Dependabot configuration to daily checks
This commit changes the Dependabot configuration to perform dependency checks daily instead of weekly. This adjustment ensures that the project stays up-to-date with the latest dependency versions more promptly, allowing for quicker identification and resolution of potential vulnerabilities or bugs in third-party packages. By increasing the frequency of checks, the project can maintain higher security and reliability standards, benefiting from the latest improvements and fixes in its dependencies.