[build] Override Spring Framework version in pom.xml
This commit adds `<spring-framework.version>6.1.4</spring-framework.version>` to the properties section of the pom.xml file. This change explicitly sets the Spring Framework version to 6.1.4, overriding the version provided by the Spring Boot Bill of Materials (BOM). This update ensures that the project uses the specified version of the Spring Framework, allowing for greater control over the dependencies and compatibility with the project's specific requirements.