[deps] Update Spring Framework version from 6.1.4 to 6.1.5
This commit updates the Spring Framework version from 6.1.4 to 6.1.5 in the pom.xml. This upgrade addresses the latest bug fixes, performance improvements, and security patches provided in the 6.1.5 release. By keeping up-to-date with the Spring Framework versions, the project ensures it benefits from the latest enhancements and maintains compatibility with the evolving Spring ecosystem.