[ci] Modify Release Workflow to trigger on pom.xml changes and check version
This commit updates the Release Workflow in GitHub Actions to trigger only on pushes to the `main` branch that include modifications to `pom.xml`. The workflow now contains a `check-version` job that compares the version in `pom.xml` against the latest git tag to determine if there has been a version update. If no version change is detected, the job exits successfully, preventing unnecessary release attempts. The `release` job is configured to run only if the `check-version` job indicates a version change, ensuring that releases are only created when necessary. This optimization of the workflow improves the efficiency and accuracy of the release process, aligning it more closely with project changes.